# school-results-db/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "School Results & Examination API",
    "description": "Complete backend API service for Examination, Marks Management, Lifecycles, and Frozen Report Cards",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/results/class-sheet": {
      "get": {
        "summary": "Get class result sheet with ranks",
        "tags": [
          "Results & Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "examClassId",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "sectionId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/results/subject-stats": {
      "get": {
        "summary": "Get subject statistics from materialized view",
        "tags": [
          "Results & Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "examClassId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/results/annual/compute": {
      "post": {
        "summary": "Compute cumulative annual results from configured weights",
        "tags": [
          "Results & Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "configId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}